Update taxonomy test
